Handle delete requests for teachers and classrooms in case of success and error. Read axios, vue and bulma from public directory instead of using them via CDN

+++ app/Http/Controllers/ClassroomsController.php
@@ -31,7 +31,11 @@ class ClassroomsController extends Controller
 
     public function destroy(Classroom $classroom)
     {
-      $classroom->delete();
-      return ['message' => 'Aula Eliminada!', 'object' => $classroom];
+      if(!$classroom->hasWorkshops()) {
+        $classroom->delete();
+        return ['message' => 'Aula Eliminada!', 'object' => $classroom];
+      } else {
+        return response()->json(['message' => 'No se pudo eliminar el aula!'], 422);
+      }
     }
 }
